refactor(display): split output/display.py by output surface
Move the 880-line output/display.py into focused per-surface modules:

- output/_console.py        — canonical Console() instance
- output/formatting.py      — byte / param / date / color helpers
- output/ranking.py         — display_ranking + display_hardware
- output/plan.py            — display_plan
- output/upgrade.py         — display_upgrade + summarize/verdict helpers
- output/json_output.py     — display_json + display_plan_json + display_upgrade_json
- output/display.py         — thin re-export shim (38 lines, was 880)

Tests that captured output by patching display.console now patch
whichllm.output._console.console instead, which is the new single source
of truth that every surface module looks up at call time.

Behavior-preserving: no ranking, plan, upgrade, or JSON output changes.
All 220 tests pass.

Refs #41
